wip: implements index file on s3 and clean up tests

+++ replibyte/src/bridge/mod.rs
@@ -24,5 +24,5 @@ pub struct IndexFile {
 pub struct Backup {
     directory_name: String,
     size: usize,
-    created_at: u64,
+    created_at: u128,
 }
